More reaction functionality

+++ app/Http/Controllers/ideaControl.php
@@ -87,19 +87,24 @@ public function comment(Idea $idea, Request $request){
 public function like(Idea $idea){
    $check_id=$idea->reaction->where('user_id',auth()->id());
   dd(count($check_id));
-/*   if(count($check_id)==1){
-    $idea->reaction->where('user_id', auth()->id())->delete();
+
+   if(count($check_id)==1){
+       $idea->reaction->where('user_id', auth()->id())->delete();
     $idea->upvote-=1;
     $idea->save();
     return back()->with('unliked');
-}*/
+   }
+
     $data=[
         "user_id"=>auth()->id(),
         "idea_id"=>$idea->id 
 ];
+$react=reaction::create($data);
+
     $idea->upvote+=1;
     $idea->save();
     return back()->with('liked');
+
 }
 
 }
